feat: big casts refactoring
* Replaced "to" cast functions with unified "from" notation
* Marked cast functions as PARALLEL SAFE to allow parallel queries
* Cast functions that cannot error marked as LEAKPROOF (may improve performance when using Postgres RLS, see https://dba.stackexchange.com/a/210391/160775)
* Correctly defined IMPLICIT vs ASSIGNMENT casts:
  * IMPLICIT for error-free casts
  * ASSIGNMENT for casts that may error
* Native casts for/from float4/float8 types

+++ sql/test_uint8.sql
@@ -222,26 +222,26 @@ SELECT s FROM generate_series(1::uint8, 10::uint8, 2::uint8) s;
 
 -- Ranges block
 
-SELECT uint8range(0, 10);
+SELECT uint8range(0::uint8, 10::uint8);
 SELECT uint8range(0::uint8, 18446744073709551615::uint8);
 SELECT uint8range(0::uint8, 18446744073709551615::uint8, '[]');
-SELECT upper(uint8range(0, 10));
-SELECT lower(uint8range(0, 10));
-SELECT isempty(uint8range(0, 10));
-SELECT uint8range(0, 10) @> 9::uint8;
-SELECT uint8range(0, 10) @> 10::uint8;
-SELECT uint8range(0, 10) && uint8range(10,20);
-SELECT uint8range(0, 10) && uint8range(9,20);
-SELECT uint8range(5, 10) - uint8range(5, 10);
-SELECT uint8range(5, 10) - uint8range(5, 9);
+SELECT upper(uint8range(0::uint8, 10::uint8));
+SELECT lower(uint8range(0::uint8, 10::uint8));
+SELECT isempty(uint8range(0::uint8, 10::uint8));
+SELECT uint8range(0::uint8, 10::uint8) @> 9::uint8;
+SELECT uint8range(0::uint8, 10::uint8) @> 10::uint8;
+SELECT uint8range(0::uint8, 10::uint8) && uint8range(10::uint8,20::uint8);
+SELECT uint8range(0::uint8, 10::uint8) && uint8range(9::uint8,20::uint8);
+SELECT uint8range(5::uint8, 10::uint8) - uint8range(5::uint8, 10::uint8);
+SELECT uint8range(5::uint8, 10::uint8) - uint8range(5::uint8, 9::uint8);
 CREATE TEMPORARY TABLE test_uint8range (
     r uint8range,
 
     EXCLUDE USING GIST (r WITH &&)
 );
 
-INSERT INTO test_uint8range (r) VALUES (uint8range(0, 10));
-INSERT INTO test_uint8range (r) VALUES (uint8range(10, 20));
-INSERT INTO test_uint8range (r) VALUES (uint8range(19, 30));
+INSERT INTO test_uint8range (r) VALUES (uint8range(0::uint8, 10::uint8));
+INSERT INTO test_uint8range (r) VALUES (uint8range(10::uint8, 20::uint8));
+INSERT INTO test_uint8range (r) VALUES (uint8range(19::uint8, 30::uint8));
 
 DROP TABLE test_uint8range;
